% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ens_mean_and_sd}
\alias{ens_mean_and_sd}
\title{Compute the ensemble mean and standard deviation}
\usage{
ens_mean_and_sd(ens)
}
\arguments{
\item{ens}{A 3d array with the third dimension as the ensemnble member}
}
\value{
A list with \code{ens_sd}, the ensemble standard deviation and
  \code{ens_mean}, the ensemble mean
}
\description{
Compute the ensemble mean and standard deviation
}
\examples{
my_file <- get_data_filename(20180101, 0)
t2m <- read_members(my_file$filepath, my_file$filename, 9, "T2m", lead_time = 0)
my_mean_and_sd <- ens_mean_and_sd(t2m$model_data)
}

