% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_mean_and_sd.R
\name{ensemble_mean}
\alias{ensemble_mean}
\title{Compute the ensemble mean of a 2d field}
\usage{
ensemble_mean(.ens)
}
\arguments{
\item{.ens}{ensemble data as read in by \code{read_members}.}
}
\value{
a list with the same format as the input but the \code{model_data}
  element is the ensemble mean.
}
\description{
Compute the ensemble mean of a 2d field
}
