% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_mean_and_sd.R
\name{ensemble_sd}
\alias{ensemble_sd}
\title{Compute the ensemble standard deviation of a 2d field}
\usage{
ensemble_sd(.ens)
}
\arguments{
\item{.ens}{Ensemble data as read in by \code{read_members}.}
}
\value{
A list with the same format as the input but the \code{model_data}
  element is the ensemble mean.
}
\description{
Compute the ensemble standard deviation of a 2d field
}

