% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_filename.R
\name{get_data_filename}
\alias{get_data_filename}
\title{Gets the filename of a forecast file}
\usage{
get_data_filename(data_date, data_cycle, data_source = "operational meps",
  experiment_name = NULL, parameter = NULL, lead_time = NULL)
}
\arguments{
\item{data_date}{The required forecast date. Maybe YYYYMMDD or YYYY-MM-DD or
any date object.}

\item{data_cycle}{The forecast cycle. Can be numeric or a string.}

\item{data_source}{If "operational meps" (the default) will find MEPS data.
Otherwise an experiment is assumed and the path must be given. Not yet
implemented.}

\item{experiment_name}{The experiment name.}

\item{parameter}{If the parameter is in a surfex file, for example, it must
be given so that the correct file is found. Otherwise an extracted file is
assumed.}

\item{lead_time}{The lead time must be given in hours for experiments.}
}
\value{
A list containting \code{\strong{fullpath}}, the full path to the
  file, \code{\strong{filepath}}, the path to the data file and
  \code{\strong{filename}}, the name of the file.
}
\description{
Works for Operational MEPS. For experments, the experiment directory is
currently hard coded. This needs fixing!
}
\examples{
get_data_filename(20180101, 0)
get_data_filename(20180101, 0, parameter = "SST")
get_data_filename("2018-01-01", "06")
get_data_filename(20180101, 0, data_source = "experiment", experiment_name = "my_exp", lead_time = 24)
}

