% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_x_section.R
\name{plot_x_section}
\alias{plot_x_section}
\title{Plot a vertical cross section of a three dimensional field}
\usage{
plot_x_section(.fcst_3d, point1, point2, surface_pressure,
  .proj4_string = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  topo_color = "grey30", topo_highlight = "grey70", legend = TRUE,
  pressure_res = 10, horizontal_res = 1, ...)
}
\arguments{
\item{.fcst_3d}{A three dimseninal mepsr type object. Must have z as a named
list element containng the level numbers.}

\item{point1}{A vector of length 2 with the (x, y) location of the first
point in the cross section. x and y will normally be longitude and latitude
repsectively in decimal degrees. If x and y are in projected co-ordinates
the proj4 string must be supplied in the \code{.proj4_string} argument.}

\item{point2}{A vector of length 2 with the (x, y) location of the second
point in the cross section. x and y must be in the same projection as
\code{point1}.}

\item{surface_pressure}{A two dimensional mepsr type object with the surface
pressure in Pa on the same domain as \code{.fcst_3d}.}

\item{.proj4_string}{The proj4 string for the cross section end points.}

\item{topo_color}{The colour to shade the topogrpahy.}

\item{topo_highlight}{The colour to highlight the top of the topography.}

\item{legend}{Whether to plot a legend (TRUE / FALSE)}

\item{pressure_res}{Vertical resolution of pressure levels in cross section
(hPa).}

\item{horizontal_res}{Horizontal resolution of cross section (km)}

\item{...}{Arguments to \link[graphics]{image} and \link[fields]{image.plot}.}
}
\description{
Plot a vertical cross section of a three dimensional field
}
