% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plot_field.R
\name{quick_plot_field}
\alias{quick_plot_field}
\title{Plot a field from an eps model}
\usage{
quick_plot_field(eps_field, member, legend = TRUE, hires_coast = FALSE, ...)
}
\arguments{
\item{eps_field}{Output of \code{read_members}.}

\item{member}{Member to plot. Can be numeric or one of "mean", "sd".}

\item{legend}{Whether to plot a legend - TRUE or FALSE. Defaults to TRUE.}

\item{hires_coast}{Plot a high resolution coastline - TRUE or FALSE. Defaults
to FALSE.}

\item{...}{Arguments for \code{image.plot}, e.g. for breaks and colour
palette.}
}
\value{
A plot
}
\description{
Uses the output of \code{read_members} to quickly plot a field with a
coastline map
}
\examples{
my_file <- get_data_filename(20180101, 0)
t2m <- read_members(my_file$filepath, my_file$filename, 9, "T2m", lead_time = 0)
quick_plot_field(t2m, 0)

breaks <- c(-50, -40, - 30, seq(-20, 15, 5))
colour_palette <- c("blue", "skyblue", "lightgreen", "green", "yellow", "gold")
colours <- colorRampPalette(colour_palette)(length(breaks) - 1)
quick_plot_field(t2m, 0, col = colours, breaks = breaks, lab.breaks = breaks)

library(viridis)
quick_plot_field(t2m, 0, hires = TRUE, col = viridis(256))
}

