% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_members.R
\name{read_members}
\alias{read_members}
\title{Read ensemble members from a forecast file}
\usage{
read_members(filepath, filename, num_perturbed_members, parameter,
  file_type = NULL, lead_time = NULL, ...)
}
\arguments{
\item{filepath}{The path to the forecast file.}

\item{filename}{The forecast file name.}

\item{num_perturbed_members}{The number of perturbed members in the forecast.
Note that this does not include the control member.}

\item{parameter}{The parameter to read.}

\item{file_type}{The forecast file format. The function can attempt to
ascertain the format from the file name, but if it can't \code{file_type}
must be passed as an argument.}

\item{lead_time}{The lead time to read.}

\item{...}{Arguments to be passed to \code{read_netcf} or \code{read_grib}}
}
\value{
A list containing: \cr
\code{\strong{model_data}}: The 3d field (the 3rd dimension is time). \cr
\code{\strong{x}}: The x coordinates in the projection of the forecast file. \cr
\code{\strong{y}}: The y coordinates in the projection of the forecast file. \cr
\code{\strong{proj4_string}}: The proj4 projection string for the forecast file. \cr
\code{\strong{parameter}}: The parameter that the 3d field represents. \cr
\code{\strong{filename}}: The full path to the forecast file.
}
\description{
\code{read_members} reads all esemble members from a forecast file for a
single lead time. It is possible to read from grib or netcdf files via the
helper functions. It is assumed that grib files only contain one member and
ntcdf files contain all members.
}
\examples{
file_path <- "/lustre/storeB/immutable/archive/projects/metproduction/MEPS/2018/01/01"
file_name <- "meps_extracted_2_5km_20180101T00Z.nc"
model_field <- read_members(file_path, file_name, 9, "precipitation_amount_acc", lead_time = 24)
model_field <- read_members(file_path, file_name, 9, "Pcp", lead_time = 24)
}

